+++ docs/CHANGE_LOG.md
@@ -1,5 +1,14 @@
 # Calendar.js - Change Log:
 
+## Version 2.3.3:
+- Added working hours support! This is shown as a slightly darker area on the Full Day view.
+- Added new setting "workingHoursStart" (defaults to null), which states the time the working hours start (for example, "09:00").
+- Added new setting "workingHoursEnd" (defaults to null), which states the time the working hours end (for example, "17:00").
+- Added dark-mode CSS support for :root variables (in light-mode, a slightly lighter dark color is now used).
+- Fixed documentation in the README.md file.
+
+<br>
+
 ## Version 2.3.2:
 - When pressing escape, if events are selected and another view is open, the events selected are cleared as the first action (the view will remain open).
 - When pressing escape, if the "Search" dialogs search history drop down is open, it is closed (instead of closing the dialog itself).